90% deploy

+++ product/models.py
@@ -30,7 +30,7 @@ class Product(models.Model):
     category = models.ForeignKey(Category, on_delete=models.CASCADE)
     name = models.CharField(max_length=100)
     image = models.ImageField(upload_to = 'images/products/%Y-%m-%d/')
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=19, decimal_places=3)
     amount = models.IntegerField()
     avg_rating = models.IntegerField(null=True, blank=True, default=0)
     description = models.TextField(blank=True, null=True)
@@ -38,7 +38,7 @@ class Product(models.Model):
     create_at = models.DateTimeField(auto_now_add=True)
     update_at = models.DateTimeField(auto_now=True)
     have_discount = models.BooleanField(default=False)
-    discount_price = models.FloatField(blank=True, null=True)
+    discount_price = models.DecimalField(max_digits=19, decimal_places=3,blank=True, null=True)
 
     def save(self, *args, **kwargs):
         super().save(*args, **kwargs)
